requirements.txt
# Core
diffusers==0.33.1
accelerate==1.6.0
transformers==4.51.3
optimum.quanto==0.2.7
peft==0.17.1

# Math / Sci
scipy==1.15.2
einops==0.8.1
einops-exts==0.0.4
setuptools==81.0.0

# Vision
opencv-python==4.11.0.86
timm==1.0.17
kornia==0.8.2
open_clip_torch==3.2.0
lpips==0.1.4
scikit-image==0.25.2
dreamsim==0.2.1
# image-reward==1.5
openai-clip==1.0.1

# NLP / tokenizer
sentencepiece==0.2.1
protobuf==6.30.2
